moved to dump_config
